The refactor for guns is finished; create new gun flamethrower
