+++ MFGJ-2021-January/Assets/Prefabs/Bullets/Player_Ammo/FireBall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 15c5effebb60c8e4aa03f381f67938ec
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
