+++ MFGJ-2021-January/Assets/Prefabs/FactoryGuns/AssaultRifle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ea92ca37beb84404ab5402d2b02e9ea0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
