+++ MFGJ-2021-January/Assets/Prefabs/FactoryGuns/Flamethrower.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 78e09ff130e304a40b70aef966cfac50
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
