+++ MFGJ-2021-January/Assets/Prefabs/FactoryGuns/Guns Configuration.asset
@@ -14,3 +14,5 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   guns:
   - {fileID: 7938697487079501078, guid: 4ca552410b6847945ba61bf3a5e8c4fe, type: 3}
+  - {fileID: 527775499845344509, guid: ea92ca37beb84404ab5402d2b02e9ea0, type: 3}
+  - {fileID: 5602091609524199567, guid: 78e09ff130e304a40b70aef966cfac50, type: 3}
