+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76859e09fa8bd3b4ab77f3eae821f1fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
