+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/FlamethrowerGun.cs
@@ -1,0 +1,3 @@
+﻿public class FlamethrowerGun : Gun
+{
+}
