+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/FlamethrowerGun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 618e0dce8b0f0414ab5b967437ca5b1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
