+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/GunsCollection.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4de82831f237aae4abddacbe841ce6e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
