+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/GunsCollection/AssaultRifle.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AssaultRifle : Gun
+{
+}
