+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/GunsCollection/AssaultRifle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a7afa7006cf74c48a213cb801feebaa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
