+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/GunsFactory.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+
+public class GunsFactory
+{
+    private readonly GunsConfiguration gunsConfiguration;
+
+    public GunsFactory(GunsConfiguration gunssConfiguration)
+    {
+        this.gunsConfiguration = gunssConfiguration;
+    }
+        
+    public Gun Create(string id)
+    {
+        var prefab = gunsConfiguration.GetGunPrefabById(id);
+
+        return Object.Instantiate(prefab);
+    }
+}
