+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/GunsFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ae2c0f3ce1d2dd44ac7fbdc04342f4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
