+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/IGun.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+public interface IGun
+{
+    void Shoot(GameObject shotPoint);
+}
