+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/IGunnig.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+public interface IGunnig
+{
+    string GetFirstGun();
+    GameObject GetShootPoint();
+}
