+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/Pistol.cs
@@ -1,0 +1,5 @@
+﻿using UnityEngine;
+
+public class Pistol : Gun
+{
+}
