+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/Pistol.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f2ecf19375a54a70897fa575c28a7aca
+timeCreated: 1619985002
