+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85570d926e0ac4f41841834f2f77a69d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
