+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/ViewGuns.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ded78eebef00a743b6135b18cce05b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
