+++ MFGJ-2021-January/Assets/Scripts/Player/Bellseboss/ViewGuns/GunsReferenceForEquip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b3b5d5423c326e49aeac8dff234108a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
