--- MFGJ-2021-January/Assets/Scripts/Player/Gun.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1da28d73c4d549bd9d78c55aac782d74
-timeCreated: 1619984225
