--- MFGJ-2021-January/Assets/Scripts/Player/GunsConfiguration.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ba8170c7b45d42cfbc0412eb6bf6931c
-timeCreated: 1619984301
