--- MFGJ-2021-January/Assets/Scripts/Player/IGun.cs
@@ -1,6 +1,0 @@
-﻿using UnityEngine;
-
-public interface IGun
-{
-    void Shoot(GameObject shotPoint);
-}
