--- MFGJ-2021-January/Assets/Scripts/Player/IGun.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ace03385e18a400ea3dc45d001098957
-timeCreated: 1619984177
