--- MFGJ-2021-January/Assets/Scripts/Player/IGunnig.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-public interface IGunnig
-{
-    string GetFirstGun();
-    GameObject GetShootPoint();
-}
