--- MFGJ-2021-January/Assets/Scripts/Player/IGunnig.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d479bfca54c342d08cf724c4baf751ef
-timeCreated: 1619983683
