--- MFGJ-2021-January/Assets/Scripts/Player/Pistol.cs
@@ -1,4 +1,0 @@
-﻿public class Pistol : Gun
-{
-    
-}
