--- MFGJ-2021-January/Assets/Scripts/Player/ShootingAndToEquipGuns.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f983a43725cc4351b9c90438314bb91c
-timeCreated: 1619983524
